% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{award}
\alias{award}
\title{Player award history for the National Women's Soccer League (NWSL)}
\format{A data frame with one row for each award in each season awarded to an NWSL player
\describe{
\item{person_id}{the id of the player}
\item{season}{season in which the award was won}
\item{award}{name of the award}
}}
\usage{
award
}
\description{
Player award history for the National Women's Soccer League (NWSL)
}
\keyword{datasets}
