% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fieldplayer_overall_season_stats}
\alias{fieldplayer_overall_season_stats}
\title{Player stats for field players in each season}
\format{
A data frame with one row for each player in each season for which data is available (2013-2019) with statistics for field players
\describe{
\item{person_id}{the id of the player beginning at 1}
\item{season}{season for which the statistics represent}
\item{nation}{nationality of the player}
\item{pos}{position}
\item{team_id}{the id of team as they appear across all datasets in \code{nwslR}}
\item{mp}{total number of matches the player appeared in}
\item{starts}{total number of matches the player started}
\item{min}{total number of minutes played}
\item{gls}{goals scored}
\item{ast}{assists}
\item{pk}{penalty kicks made}
\item{p_katt}{penalty kicks attempted}
\item{crd_y}{total number of yellow cards received}
\item{crd_r}{total number of red cards received}
}
}
\source{
\url{https://fbref.com/en/comps/182/keepers/NWSL-Stats}
}
\usage{
fieldplayer_overall_season_stats
}
\description{
Player stats for field players in each season
}
\keyword{datasets}
