% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_adv_player_stats.R
\name{get_adv_player_stats}
\alias{get_adv_player_stats}
\title{Advanced player-level statistics for each game}
\usage{
get_adv_player_stats(game_id)
}
\arguments{
\item{game_id}{Unique game id from nwsl. Find these on the NWSL website (ex: https://www.nwslsoccer.com/game/washington-spirit-vs-sky-blue-2019-04-13) or in the \code{game} table.}
}
\description{
Scrapes NWSL website to pull player stats for each game. Returns a data frame with one row per player on each team
game ID (home and away) as well as over 200 different statistics
}
\details{
Most variables are documented at the following link, but official documentation will be added to the pacakge soon: https://winsports.dayscript.com/docs/database/match_player_stats
}
