% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_adv_team_stats.R
\name{get_adv_team_stats}
\alias{get_adv_team_stats}
\title{Advanced team-level statistics by game}
\usage{
get_adv_team_stats(game_id)
}
\arguments{
\item{game_id}{Unique game id from nwsl. Find these on the NWSL website (ex: https://www.nwslsoccer.com/game/washington-spirit-vs-sky-blue-2019-04-13) or in the \code{game} table.}
}
\description{
Scrapes NWSL website to pull team stats for each game. Returns a data frame with two rows for each
game ID (home and away) as well as up to 200 different statistics depending on the game
}
