% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_matches.R
\name{load_matches}
\alias{load_matches}
\title{Loads match table}
\usage{
load_matches()
}
\description{
Loads match table that holds all matches played between 2016-present
w/ match IDs
}
\note{
Most, but not all, matches in table have associated stats that you can
pull by feeding match ID into \code{load_player_match_stats} or
\code{load_team_match_stats}. Match start times are in East Coast USA time.
}
