% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_metrics.R
\name{load_metrics}
\alias{load_metrics}
\title{Metric definitions}
\usage{
load_metrics()
}
\description{
Data dictionary holding all definitions for metrics pulls from
\code{load_player_match_stats}, \code{load_team_match_stats},
\code{load_player_season_stats}, \code{load_team_season_stats}
}
\note{
Not all metrics all available for all players, matches, teams, or seasons
}
