% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_match_stats.R
\name{load_player_match_stats}
\alias{load_player_match_stats}
\title{Loads player level match stats}
\source{
\url{www.nwslsoccer.com}
}
\usage{
load_player_match_stats(match_id = c())
}
\arguments{
\item{match_id}{match id available from \code{load_matches()}}
}
\description{
Loads player level stats for a given match ID
}
\examples{
\donttest{
match <- load_player_match_stats("portland-thorns-fc-vs-kansas-city-current-2022-10-29")
}

}
\seealso{
See \code{load_metrics} for information about returned metrics. Not all
metrics exist for a given match. Join on \code{load_players()} and \code{load_teams()}.
}
