% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_season_stats.R
\name{load_team_season_stats}
\alias{load_team_season_stats}
\title{Loads team level season stats}
\source{
\url{www.nwslsoccer.com}
}
\usage{
load_team_season_stats(team_id, season)
}
\arguments{
\item{team_id}{team_id available from \code{available_teams()}}

\item{season}{season available from \code{available_seasons()}}
}
\description{
Loads team level season stats for a given team ID/season
}
\examples{
\donttest{
season <- load_team_season_stats(team_id = "POR", season = "2022")
short_season <- load_team_season_stats(team_id = "POR", season = "2022_NWSL_Challenge_Cup")
}

}
\seealso{
See \code{load_metrics} for information about returned metrics. Not all
metrics exist for a given match. Join on \code{load_teams()}.
}
