% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{read_nwsl_data}
\alias{read_nwsl_data}
\title{Provides user-friendly error messages}
\usage{
read_nwsl_data(url)
}
\arguments{
\item{url}{url to pull file from}
}
\description{
Wrapper on \code{readr::read_csv()} to create better error messages when
match data does not exist
}
