% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stadium}
\alias{stadium}
\title{Stadium information for games played each season}
\format{
A data frame with one row for each team in each season for all seasons with information regarding home stadiums and average attendance
\describe{
\item{team_id}{the id of team as they appear across all datasets in \code{nwslR}}
\item{pri_stadium_name}{name of primary home stadium}
\item{pri_stadium_alias}{if stadium was ever known by another name during the existence of the team, it is denoted here}
\item{pri_town}{town of primary home stadium}
\item{pri_state}{state of primary home stadium}
\item{pri_capacity}{seating capacity of primary home stadium}
\item{sec_stadium_name}{name of secondary home stadium (only if home games were ever played elsewhere during the season)}
\item{sec_stadium_alias}{if stadium was ever known by another name during the existence of the team, it is denoted here}
\item{sec_town}{town of secondary home stadium}
\item{sec_state}{state of secondary home stadium}
\item{sec_capacity}{seating capacity of secondary home stadium}
\item{avg_attendance}{average season home game attendance}
\item{season}{season for which the statistics represent}
}
}
\source{
\url{https://en.wikipedia.org/wiki/List_of_National_Women\%27s_Soccer_League_stadiums}
}
\usage{
stadium
}
\description{
Stadium information for games played each season
}
\keyword{datasets}
