% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{team_stats_season}
\alias{team_stats_season}
\title{Team stats for each season}
\format{A data frame with one row for each team in each season for which data is available (2016-2019) with statistics for team performance
\describe{
\item{team_id}{the id of team as they appear across all datasets in \code{nwslR}}
\item{goals}{total number of goals scored}
\item{assists}{total number of assists}
\item{shots}{total number of shots}
\item{shots_on_goals}{total number of shot on goal}
\item{fouls_committed}{total number of fouls committed by team}
\item{fouls_suffered}{total number of time the team was fouled}
\item{offsides}{total number of offside calls}
\item{corner_kicks}{total number of corner kicks taken}
\item{games_played}{total number of games played}
\item{yellow_cards}{total number of yellow cards received}
\item{red_cards}{total number of red cards received}
\item{penalty_kick_goals}{total number of goals scored from penalty kicks}
\item{penality_kick_attempts}{total number of penalty kicks awarded}
\item{season}{season for which the statistics represent}
\item{wins}{number of games won}
\item{draws}{number of games drawn}
\item{losses}{number of games lost}
}}
\source{
\url{http://www.nwslsoccer.com/stats#teams}
}
\usage{
team_stats_season
}
\description{
Team stats for each season
}
\keyword{datasets}
