% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{draftpicks}
\alias{draftpicks}
\title{National Women's Soccer League (NWSL) Draft Picks since 2013}
\format{A data frame with one row per player selected}
\source{
Wikipedia (16 Jan 2020)
\describe{
\item{year}{Year of the draft}
\item{round}{Round of the draft}
\item{pick}{Overall selection number}
\item{nation}{Player's home country}
\item{player}{player name}
\item{pos}{position(s) declared}
\item{team_name}{team making the pick}
\item{previous_team}{college and/or previous club side}
\item{notes}{additional notes, indicating a pick was obtained via trade or other means}
\item{person_id}{the id of the player}
\item{team_id}{the id of team as they appear across all datasets in \code{nwslR}}
}
}
\usage{
draftpicks
}
\description{
National Women's Soccer League (NWSL) Draft Picks since 2013
}
\keyword{datasets}
