% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adv_team_stats.R
\name{get_adv_team_stats}
\alias{get_adv_team_stats}
\title{Scrapes NWSL website to pull team stats for each game. Returns a data frame with two rows for each
game ID (home and away) as well as over 200 different statistics}
\usage{
get_adv_team_stats(year)
}
\arguments{
\item{year}{year between 2016-2019}
}
\description{
Scrapes NWSL website to pull team stats for each game. Returns a data frame with two rows for each
game ID (home and away) as well as over 200 different statistics
}
